# src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "Index Card API",
    "description" : "This is a simple server for loading and saving index cards",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://dev.index-cards.com/v1",
    "description" : "Dev Server"
  }, {
    "url" : "https://prod.index-cards.com/v1",
    "description" : "Prod Server"
  } ],
...
  "paths" : {
...
    "/index-cards/{userId}" : {
      "get" : {
        "tags" : [ "index-cards" ],
        "summary" : "get all cards of the logged in user",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "9666e8bf-6f37-44d1-aab4-f2faaf6da243"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successfully pulled the users cards",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Index-card"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/index-cards/{cardId}" : {
      "get" : {
        "tags" : [ "index-cards" ],
        "summary" : "Get a specific card from the database",
        "parameters" : [ {
          "name" : "cardId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "9666e8bf-6f37-44d1-aab4-f2faaf6da243"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully pulled the requested card",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Index-card"
                }
              }
            }
          },
          "404" : {
            "description" : "Card not found"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "index-cards" ],
        "summary" : "delete specified card",
        "parameters" : [ {
          "name" : "cardId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "9666e8bf-6f37-44d1-aab4-f2faaf6da243"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Successfully deleted card"
          },
          "404" : {
            "description" : "Card not found"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "patch" : {
        "tags" : [ "index-cards" ],
        "summary" : "update specified card",
        "parameters" : [ {
          "name" : "cardId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "9666e8bf-6f37-44d1-aab4-f2faaf6da243"
          }
        } ],
        "requestBody" : {
          "description" : "New information to update card",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Index-card"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "successfully updated selected card",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Index-card"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/index-cards/{collectionId}/index-cards" : {
      "get" : {
        "tags" : [ "index-cards" ],
        "summary" : "get all cards from the collection",
        "parameters" : [ {
          "name" : "collectionId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "9666e8bf-6f37-44d1-aab4-f2faaf6da243"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully pulled all cards from collection",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Index-card"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "index-cards" ],
        "summary" : "Create a new index card for the selected collection",
        "parameters" : [ {
          "name" : "collectionId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "9666e8bf-6f37-44d1-aab4-f2faaf6da243"
          }
        } ],
        "requestBody" : {
          "description" : "content of the index card",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Index-card"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "successfully created card",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Index-card"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Index-card" : {
        "type" : "object",
        "properties" : {
          "cardId" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "9666e8bf-6f37-44d1-aab4-f2faaf6da243"
          },
          "collectionId" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "9666e8bf-6f37-44d1-aab4-f2faaf6da243"
          },
          "frontSide" : {
            "type" : "string",
            "example" : "What is the capital of Portugal?"
          },
          "frontFormat" : {
            "type" : "string",
            "example" : "latin",
            "enum" : [ "latin", "hiragana", "katakana", "kanji" ]
          },
          "backSide" : {
            "type" : "string",
            "example" : "Lisbon"
          },
          "backFormat" : {
            "type" : "string",
            "example" : "latin",
            "enum" : [ "latin", "hiragana", "katakana", "kanji" ]
          }
        },
        "description" : "Model for the index cards",
        "example" : {
          "backFormat" : "latin",
          "backSide" : "Lisbon",
          "cardId" : "9666e8bf-6f37-44d1-aab4-f2faaf6da243",
          "frontFormat" : "latin",
          "frontSide" : "What is the capital of Portugal?",
          "collectionId" : "9666e8bf-6f37-44d1-aab4-f2faaf6da243"
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}